+++ syft/format/decoders_collection_test.go
